% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apdesign.R
\docType{package}
\name{apdesign}
\alias{apdesign}
\alias{apdesign-package}
\title{apdesign: AP design matrix for time-unstructured data}
\usage{
apdesign(data, id_var, time_var, center_time, cycle_var, center_cycle,
  max_degree = c(1, 1))
}
\arguments{
\item{data}{A data frame.}

\item{id_var}{A character that indicates the subject identifier in 
\code{data}.}

\item{time_var}{A character that indicates the within-cycle time indicator in 
\code{data}.}

\item{center_time}{A numeric specifying the within-cycle time to center on.}

\item{cycle_var}{A character that indicates the cycle indicator in 
\code{data}.}

\item{center_cycle}{A numeric specifying the cycle to center on.}

\item{max_degree}{A vector of numerics specifying the highest degree for 
each polynomial.}
}
\value{
Output will be a data frame.
}
\description{
This is an implementation of the AP coding for varying numbers of measurement 
occasions and/or variably spaced measurement occasions.

apdesign: AP coding
\code{apdesign} returns a data frame with additive polynomial coding
}

