% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_apdesign_i.R
\name{apdesign_i}
\alias{apdesign_i}
\title{apdesign_i: AP coding for a single subject}
\usage{
apdesign_i(data, cycle_var, center_cycle, time_var, center_time,
  max_degree = c(1, 1), matricies = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{cycle_var}{A character that indicates the cycle indicator in 
\code{data}.}

\item{center_cycle}{A numeric specifying the cycle to center on.}

\item{time_var}{A character that indicates the within-cycle time indicator in 
\code{data}.}

\item{center_time}{A numeric specifying the within-cycle time to center on.}

\item{max_degree}{A vector of numerics specifying the highest degree for 
each polynomial.}

\item{matricies}{If \code{TRUE}, will print the AP, D1 and D2 matricies.}
}
\value{
Output will be a matrix.
}
\description{
apdesign_i: AP coding for a single subject
}

