% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{three_season_cycle}
\alias{three_season_cycle}
\title{Longitudinal test scores over three school years.}
\format{A data frame with 234 rows and 6 variables:
\describe{
  \item{id}{student identifier}
  \item{cycle}{cycle number}
  \item{grd}{grade level, corresponds to grade}
  \item{season}{season within cycle/grade}
  \item{week}{week since start of the cycle}
  \item{score}{test score}
}}
\usage{
three_season_cycle
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}

