% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_apdesign.R, R/fun_apdesign_i.R
\name{apdesign}
\alias{apdesign}
\title{apdesign: AP coding
\code{apdesign} returns a data frame with additive polynomial coding}
\usage{
apdesign(data, id_var, time_var, center_time, cycle_var, center_cycle,
  max_degree = c(1, 1))

apdesign(data, id_var, time_var, center_time, cycle_var, center_cycle,
  max_degree = c(1, 1))
}
\arguments{
\item{data}{A data frame.}

\item{id_var}{A character that indicates the subject identifier in 
\code{data}.}

\item{time_var}{A character that indicates the within-cycle time indicator in 
\code{data}.}

\item{center_time}{A numeric specifying the within-cycle time to center on.}

\item{cycle_var}{A character that indicates the cycle indicator in 
\code{data}.}

\item{center_cycle}{A numeric specifying the cycle to center on.}

\item{max_degree}{A vector of numerics specifying the highest degree for 
each polynomial.}

\item{id_var}{A character that indicates the subject identifier in 
\code{data}.}
}
\value{
Output will be a data frame.

Output will be a data frame.
}
\description{
apdesign: AP coding
\code{apdesign} returns a data frame with additive polynomial coding

apdesign: AP coding
\code{apdesign} returns a data frame with additive polynomial coding
}

